backend/user/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2==2.9.10
PyJWT==2.8.0
python-dotenv==1.0.1
redis==5.2.1
SQLAlchemy==2.0.39
typing_extensions==4.12.2
Werkzeug==3.1.3
